Fixes wrong output of example05
The PC value needs to remain put when a data hazard causes the IF and ID
stages to stall. However, in example5 this behavior was **not**
observed. Instead the PC was incremented even though the IF and ID
stages were stalled. This commit fixes the issue.

Special thanks to Dong Jun Youn and Hangyeol Yu for identifying the incorrect
output.